requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
Babel==1.3
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
billiard==3.3.0.23
celery==3.1.23
certifi==14.5.14
click==6.6
configparser==3.5.0
coverage==3.7.1
decorator==4.0.6
Django==1.7.7
django-admin-honeypot==1.0.0
django-appconf==0.6
django-bower==5.0.4
django-braces==1.8.1
django-ckeditor-updated==4.4.4
django-compressor==1.6
django-debug-toolbar==1.3.0
django-dotenv==1.3.0
django-getenv==1.3.1
django-imagekit==3.2.4
django-mailviews==0.6.5
django-pipeline==1.4.7
django-smtp-ssl==1.0
django-suit==0.2.12
django-timezone-field==1.1
djangorestframework-jwt==1.11.0
DjangoSEO==1.0
djrill==1.2.0
ecdsa==0.13
enum34==1.1.6
Fabric==1.10.1
flake8==3.0.4
Flask==0.12.2
flower==0.8.2
futures==2.2.0
gnureadline==6.3.3
hiredis==0.2.0
honcho==0.7.1
httpie==0.9.9
httplib2==0.9.2
ipython==4.0.1
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.10
jsonfield==1.0.3
kombu==3.0.35
MarkupSafe==1.0
mccabe==0.5.2
mock==1.0.1
mockito==0.5.2
oauthlib==1.1.1
paramiko==1.15.2
path.py==8.1.2
pbr==1.10.0
peewee==2.8.3
pexpect==4.0.1
pickleshare==0.5
pilkit==1.1.12
Pillow==2.6.1
psycopg2==2.6
ptyprocess==0.5
pycodestyle==2.0.0
pycrypto==2.6.1
pyflakes==1.2.3
Pygments==2.2.0
PyJWT==1.6.4
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2016.4
RandomWords==0.2.0
raven==5.2.0
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.6.1
selenium==2.45.0
simplegeneric==0.8.1
simplejson==3.8.1
six==1.10.0
sorl-thumbnail==12.3
South==1.0
speedtest-cli==0.3.4
sqlparse==0.1.14
stevedore==1.17.1
tornado==4.1
trackmac==0.0.6
traitlets==4.0.0
twilio==5.4.0
URLObject==2.4.0
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.14.1
youtube-dl==2016.9.19
